+++ src/db/schema.rs
@@ -21,7 +21,7 @@ table! {
         admin -> Text,
         label -> Text,
         creation_time -> Text,
-        height -> Int8,
+        height -> Numeric,
     }
 }
 
@@ -40,8 +40,8 @@ table! {
         cw20_address -> Text,
         sender_address -> Text,
         recipient_address -> Text,
-        amount -> Int8,
-        height -> Int8,
+        amount -> Numeric,
+        height -> Numeric,
     }
 }
